fix issue with deleted files
a recent change added ":w" before all commands, to avoid losing unsaved
changes when moving to a different slide. Unfortunately, this introduced
problems when moving to a slide in which the current file no longer
exists: the ":w" re-creates the file, preventing you from moving back to
the slide in which the file still exists!

with ":wa", the file is only saved if it is modified. In particular, if
the file doesn't exist, it is left alone.