--force harder
Sometimes you want to demonstrate a command which creates a file, and
then for the next slide you have a prepared version of that generated
file, just in case the live creation failed. But then git complains that
moving to the next slide would overwrite that generated file, even with

    git-slides next --force

Now, --force will remove all unrecognized files, making sure you move to
the repository to the next slide, with the exact same state as you
committed it.