fix git-slides transform
The underlying implementation, git-rebase, changed its behaviour to open
an editor to edit the commit message when a conflict had to be resolved.
This is not useful in git-slides, where most commit messages simply say
"new slide". Thus, I set EDITOR=true so that the existing commit message
is silently accepted unchanged, like before.